+++ PSInsight/Public/Object Type Attributes/New-InsightObjectTypeAttributes.ps1
@@ -353,11 +353,11 @@ New-InsightObjectTypeAttributes -Name "Link to Parent" -Type Object -typeValue 2
         }
 
         # Add all the aditional Attributes. 
-        if ($uniqueAttribute) {
-            $RequestBody.Add("uniqueAttribute", $UniqueAttribute)
+        if ($PSBoundParameters["UniqueAttribute"]) {
+            $RequestBody.Add("UniqueAttribute", $PSBoundParameters["UniqueAttribute"])
         }
-        if ($Description) {
-            $RequestBody.Add("Description", $Description)
+        if ($PSBoundParameters["Description"]) {
+            $RequestBody.Add("Description", $PSBoundParameters["Description"])
         }
         if ($PSBoundParameters["minimumCardinality"]) {
             $RequestBody.Add("minimumCardinality", $PSBoundParameters["minimumCardinality"])